ImageValFinetune/requirements_finetune.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.41.0
accelerate>=0.20.0
peft>=0.7.0
pandas>=1.3.0
numpy>=1.21.0
tqdm>=4.62.0
Pillow>=8.0.0
openpyxl>=3.0.0
matplotlib>=3.5.0
pyyaml>=6.0
datasets>=2.0.0
deepspeed>=0.9.0